pyproject.toml
[project]
name = "restate-ai-agents"
version = "0.1.0"
description = "Python examples using Restate for agents"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "hypercorn",
    "restate_sdk==0.6.0",
    "pydantic",
    "black",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch.build.targets.wheel]
packages = ["app"]